% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdata-package.R
\docType{package}
\name{usdata-package}
\alias{usdata}
\alias{usdata-package}
\title{usdata: Data on the States and Counties of the United States}
\description{
Demographic data on the United States at the county and state levels spanning multiple years.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OpenIntroStat/usdata}
  \item Report bugs at \url{https://github.com/OpenIntroStat/usdata/issues}
}

}
\author{
\strong{Maintainer}: Mine Çetinkaya-Rundel \email{cetinkaya.mine@gmail.com} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID})

Authors:
\itemize{
  \item David Diez \email{david@openintro.org}
  \item Leah Dorazio \email{leah.dorazio@sfuhs.org}
}

}
\keyword{internal}
